move test classification data to testfixtures.py
